requirements.txt
Flask
pytesseract
Pillow
opencv-python
spacy
numpy
scikit-learn
